LayoutTests/fast/forms/input-value-sanitization-expected.txt

Tests for value sanitization algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.value is "50"
PASS successfullyParsed is true

TEST COMPLETE
